+++ Oproto.FluentDynamoDb.UnitTests/Expressions/LambdaExpressionApiGapsPropertyTests.cs
@@ -342,6 +342,9 @@ internal class TestEntityWithMetadata : IDynamoDbEntity
     public static Dictionary<string, AttributeValue> ToDynamoDb<TSelf>(TSelf entity, FluentDynamoDbOptions? options = null)
         where TSelf : IDynamoDbEntity => new();
 
+    public static Dictionary<string, AttributeValue> ToDynamoDb<TSelf>(TSelf entity, FluentDynamoDbOptions? options, KeyInputMode keyInputMode)
+        where TSelf : IDynamoDbEntity => ToDynamoDb(entity, options);
+
     public static TSelf FromDynamoDb<TSelf>(Dictionary<string, AttributeValue> item, FluentDynamoDbOptions? options = null)
         where TSelf : IReadOnlyEntity => (TSelf)(object)new TestEntityWithMetadata();