+++ Oproto.FluentDynamoDb.UnitTests/Requests/Extensions/WithClientExtensionsTests.cs
@@ -27,6 +27,9 @@ public class WithClientExtensionsTests
             };
         }
 
+        public static Dictionary<string, AttributeValue> ToDynamoDb<TSelf>(TSelf entity, FluentDynamoDbOptions? options, KeyInputMode keyInputMode)
+            where TSelf : IDynamoDbEntity => ToDynamoDb(entity, options);
+
         public static TSelf FromDynamoDb<TSelf>(Dictionary<string, AttributeValue> item, FluentDynamoDbOptions? options = null) where TSelf : IReadOnlyEntity
         {
             var entity = new TestEntity