+++ Oproto.FluentDynamoDb.UnitTests/Requests/GetItemRequestBuilderTests.cs
@@ -33,6 +33,9 @@ public class GetItemRequestBuilderTests
             };
         }
 
+        public static Dictionary<string, AttributeValue> ToDynamoDb<TSelf>(TSelf entity, FluentDynamoDbOptions? options, KeyInputMode keyInputMode)
+            where TSelf : IDynamoDbEntity => ToDynamoDb(entity, options);
+
         public static TSelf FromDynamoDb<TSelf>(Dictionary<string, AttributeValue> item, FluentDynamoDbOptions? options = null) where TSelf : IReadOnlyEntity
         {
             var entity = new TestEntity