+++ Oproto.FluentDynamoDb.UnitTests/Requests/KeyConditionBuilderTests.cs
@@ -33,6 +33,9 @@ public class KeyConditionBuilderTests
             };
         }
 
+        public static Dictionary<string, AttributeValue> ToDynamoDb<TSelf>(TSelf entity, FluentDynamoDbOptions? options, KeyInputMode keyInputMode)
+            where TSelf : IDynamoDbEntity => ToDynamoDb(entity, options);
+
         public static TSelf FromDynamoDb<TSelf>(Dictionary<string, AttributeValue> item, FluentDynamoDbOptions? options = null) where TSelf : IReadOnlyEntity
         {
             var entity = new SimpleKeyEntity
@@ -84,6 +87,9 @@ public class KeyConditionBuilderTests
                 ["sk"] = new AttributeValue { S = testEntity?.Sk ?? string.Empty }
             };
         }
+
+        public static Dictionary<string, AttributeValue> ToDynamoDb<TSelf>(TSelf entity, FluentDynamoDbOptions? options, KeyInputMode keyInputMode)
+            where TSelf : IDynamoDbEntity => ToDynamoDb(entity, options);
 
         public static TSelf FromDynamoDb<TSelf>(Dictionary<string, AttributeValue> item, FluentDynamoDbOptions? options = null) where TSelf : IReadOnlyEntity
         {