+++ Oproto.FluentDynamoDb/Requests/BatchWriteBuilder.cs
@@ -86,7 +86,7 @@ public class BatchWriteBuilder
                 if (hydrator != null && entity != null)
                 {
                     var blobProvider = options.BlobStorageProvider;
-                    var serialized = await hydrator.SerializeAsync(entity, blobProvider, options, ct).ConfigureAwait(false);
+                    var serialized = await hydrator.SerializeAsync(entity, blobProvider, options, cancellationToken: ct).ConfigureAwait(false);
                     builder.SetResolvedItem(serialized);
                     _requestItems[tableName][requestIndex].PutRequest.Item = serialized;
                 }